+++ packages/web-app/src/hooks/useProposal.tsx
@@ -1,6 +1,6 @@
 import {Address} from '@aragon/ui-components/dist/utils/addresses';
 import {useEffect, useState} from 'react';
-import {HookData} from 'utils/types';
+import {HookData, UncategorizedProposalData} from 'utils/types';
 import {useDaoProposals} from './useDaoProposals';
 
 /**
@@ -39,120 +39,3 @@ export function useProposal(
 
   return {data: proposalData, isLoading, error};
 }
-
-/* DATA STRUCTURE =========================================================== */
-
-export type ProposalData = UncategorizedProposalData & {
-  type: 'draft' | 'pending' | 'active' | 'succeeded' | 'executed' | 'defeated';
-};
-
-type Seconds = string;
-
-export type UncategorizedProposalData = {
-  id: string;
-  metadata: ProposalMetadata;
-  vote: VotingData;
-  execution: ExecutionData;
-};
-
-type ProposalMetadata = {
-  title: string;
-  description: string;
-  publisher: Address;
-  resources?: ProposalResource[];
-  published?: BlockChainInteraction;
-  executed?: BlockChainInteraction;
-};
-
-type ProposalResource = {
-  title: string;
-  url: string;
-};
-
-type BlockChainInteraction = {
-  date: Seconds;
-  block: string;
-};
-
-export type VotingData = {
-  start: Seconds;
-  end: Seconds;
-  total: number;
-  results: Record<string, number>; // e.g. option -> amount of votes
-  tokenSymbol: string;
-};
-
-type ExecutionData = {
-  from: Address;
-  to: Address;
-  amount: number;
-};
-
-/* MOCK DATA ================================================================ */
-
-const fullProposalData: UncategorizedProposalData = {
-  id: '1',
-  metadata: {
-    title: 'Proposal: Title',
-    description: 'Proposal: Description',
-    publisher: '0x1234567890123456789012345678901234567890',
-    resources: [
-      {
-        title: 'Resource title',
-        url: 'https://example.com',
-      },
-    ],
-    published: {
-      date: '1644988365',
-      block: '123456789',
-    },
-    executed: {
-      date: '1644988365',
-      block: '123456789',
-    },
-  },
-  vote: {
-    start: '1644988365',
-    end: '1644988365',
-    total: 100,
-    results: {
-      option1: 10,
-      option2: 90,
-    },
-    tokenSymbol: 'ANT',
-  },
-  execution: {
-    from: '0x1234567890123456789012345678901234567890',
-    to: '0x1234567890123456789012345678901234567890',
-    amount: 0,
-  },
-};
-const partialProposalData: UncategorizedProposalData = {
-  id: '2',
-  metadata: {
-    title: 'Proposal: Title',
-    description: 'Proposal: Description',
-    publisher: '0x1234567890123456789012345678901234567890',
-    published: {
-      date: '1644988365',
-      block: '123456789',
-    },
-  },
-  vote: {
-    start: '1644988365',
-    end: '1644988365',
-    total: 100,
-    results: {},
-    tokenSymbol: 'DAI',
-  },
-  execution: {
-    from: '0x1234567890123456789012345678901234567890',
-    to: '0x1234567890123456789012345678901234567890',
-    amount: 0,
-  },
-};
-
-const proposals: UncategorizedProposalData[] = [
-  fullProposalData,
-  partialProposalData,
-];
